[IMP] A single installation and clean up run.
Almost all container can be build in an single run, but this is not a recomended practice for containers that needs to be changed regulary. In this case this image is build only when an update is needed, added a new base packages (such things does not occurs in a regular basis).